detalles finales

+++ app/src/main/java/com/example/sportsapp/DetailActivity.kt
@@ -1,10 +1,12 @@
 package com.example.sportsapp
 
+import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.sportsapp.databinding.ActivityDetailBinding
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
+import android.widget.Toast
 
 class DetailActivity : AppCompatActivity() {
 
@@ -40,6 +42,16 @@ class DetailActivity : AppCompatActivity() {
         android.util.Log.d("FAVORITOS", "Usuario: ${user?.uid}, anon: ${user?.isAnonymous}")
 
         btn.setOnClickListener {
+            val user = FirebaseAuth.getInstance().currentUser
+
+            // Si es anónimo o no hay sesión, pedir que inicie sesión
+            if (user == null || user.isAnonymous) {
+                Toast.makeText(this, "Inicia sesión para agregar favoritos", Toast.LENGTH_SHORT).show()
+                startActivity(Intent(this, LoginActivity::class.java))
+                return@setOnClickListener
+            }
+
+            // Si está autenticado, guardar/eliminar favorito
             docRef.get().addOnSuccessListener { snap ->
                 if (snap.exists()) {
                     docRef.delete()
